add .style.yapf for formatting with yapf
yapf is a Python source code formatter, similar to clang-format. It has quite
a lot of configuration options, which are stored in the file .style.yapf.

I've adjusted the defaults from 79 to 80 columns line width, and some settins
that handle line splitting of parenthesised, comma-separated content. These
settings are my taste.

The main reason is that overly long files are not easy to read and must be
split at a certain length. This goes for comments as well. Also, git is a
line-based tool, so having long lines also leads to more merge conflicts.